requirements.txt
fastapi
uvicorn[standard]
transformers
torch
safetensors
python-dotenv
asyncpg
llama-cpp-python
transformers accelerate sentencepiece